Hard coded board
Hand adjusted the values of the board and the blocks to suit my eye. I'll need to derive a relation for later when I want to just derive it from the height/width of the board.

So, this currently draws all the spots and the three bigger variant blocks. Next I'll draw a cursor. Hopefully I'll pair that with a board object that does the above relation.