ijkl instead of arrows
Switched to ijkl keys to move the cursor instead of the arrow keys. That way, if the page is ever tall or wide, playing the game won't also move the page. The mental dissonance is just as bad as before. Perhaps if I switched (later) to holding shift or space while moving with wasd?

Also, added first test: t_apply_cursor_move(). Added license note to blurb paragraph. Updated the evolution list to include tests.