Semismart move counter
t_block.update_move_counter saves the last cursor in a global var. This one then checks if the previous cursor was in the reverse direction. So, it will subtract from the moves when in the cursor was moved from the same spot. However, if this shape is larger than a single block, moving the cursor within the shape will disrupt the current logic. Will try to fix.