Minor fn extract, game save info
Extracted the lexing functionality into its own function. Put game save keys onto the html (no use looking in the code if I forget). I may cut out the string representation of the game state in the box. It's a minor change luckily. Next is parsing, which could be a bit complex to ensure that shapes aren't violating expectations.