Minor tidy
Deleted some output statements related to debugging and switched back to pretty squares. This is likely the end of the line for this project for an indefinite future. Sounds - the incepting idea for this project - are trivial to add but unnecessary. Hint AI is a major undertaking and not conducive to my current time. I bet there'll be another in the soon future as I tweak the readme or whatever. It's been fun.