Trying dict list for github
Updated the readme with a list of where the project _can_ go in some future time. Unless this messes up the github page ( which I anticipate will happen ), This is the end of the line for this project. Well, maybe I could do a simple extraction to prototype style, but not much else.