Methods extracted to prototype
I extracted the methods from the object to their respective prototypes. I looked about and agreed that sending an object literal to the prototype reduced the pushes to the prototype object. In this case, I doubt I will see a difference. But, this about solidifies the convention I will follow henceforth.